requirements.txt
Django==5.2.11
gunicorn==25.1.0
dj-database-url==3.1.2
whitenoise==6.11.0
python-docx==1.2.0
pypdf==6.1.3
weasyprint==66.0
psycopg[binary]==3.3.3
openai>=1.0.0
numpy>=1.24.0
